finalizando

+++ Telas/TelaAdmin.py
@@ -1,4 +1,5 @@
 import tkinter as tk
+from datetime import datetime
 from socket import socket
 from tkinter import ttk, messagebox
 
@@ -21,12 +22,14 @@ class TelaAdmin(tk.Frame):
         self.disciplina_repository = DisciplinaRepository()
         self.inscricao_repository = InscricaoRepository()
 
-        self.dict_cursos = dict(self.curso_repository.get_cursos())
+        self.clicked_filter = False
 
+        self.dict_cursos = dict(self.curso_repository.get_cursos())
+        self.dict_disciplinas = dict(self.disciplina_repository.get_all_to_dict())
         self.grid_columnconfigure(0, weight=1)
         self.controller = controller
 
-        self.notebook = ttk.Notebook(self, width=1140)
+        self.notebook = ttk.Notebook(self, width=1180)
 
         self.notebook.grid(column=0, row=1, pady=30)
 
@@ -41,7 +44,6 @@ class TelaAdmin(tk.Frame):
         self.notebook.add(self.curso_frame, text="Curso")
         self.notebook.add(self.disciplina_frame, text="Disciplina")
         self.notebook.add(self.inscricoes_frame, text="Inscrições")
-
 
         ttk.Button(self, text="Sair", width=5, command=self.deslogar).place(x=1060, y=10)
 
@@ -68,13 +70,11 @@ class TelaAdmin(tk.Frame):
         ttk.Button(frame_entrys, text="Excluir", command=self.exluir_curso).place(y=200)
         colunas = [column[1] for column in self.curso_repository.get_columns_names()]
         data = self.curso_repository.get_cursos()
-        width = int(900 / len(colunas))
+        width = [300, 600]
 
         self.tree_curso = Treeview(frame_tree, colunas, data, width)
 
         self.tree_curso.bind("<<TreeviewSelect>>", self.on_curso_select)
-
-
 
 
     def draw_disciplina_frame(self):
@@ -88,8 +88,8 @@ class TelaAdmin(tk.Frame):
         self.disciplina_frame.grid_columnconfigure(1, weight=3)
         self.disciplina_frame.grid_rowconfigure(0, weight=1)
 
-        colunas = ['codigo', 'nome', 'nome_curso']
-        width = int(900 / len(colunas))
+        colunas = ['codigo', 'nome', 'nome_curso', 'ano', 'semestre']
+        width = [150, 250, 250, 90, 80]
 
         data = self.disciplina_repository.get_disciplinas()
 
@@ -103,18 +103,28 @@ class TelaAdmin(tk.Frame):
         self.codigo_entry = ttk.Entry(frame_entrys, width=20)
         self.codigo_entry.grid(row=1, column=1, pady=5)
 
+        ttk.Label(frame_entrys, text="Ano: ").grid(row=2, column=0, sticky="W")
+        self.ano_entry = ttk.Entry(frame_entrys, width=20)
+        self.ano_entry.grid(row=3, column=0, pady=5)
+
+        ttk.Label(frame_entrys, text="Semetre: ").grid(row=2, column=1, sticky="W")
+        self.semestre_entry = ttk.Entry(frame_entrys, width=20)
+        self.semestre_entry.grid(row=3, column=1, pady=5)
+
         selected_key = tk.StringVar()
 
-        ttk.Label(frame_entrys, text="Curso").grid(row=2, column=0, sticky="W")
-        self.combobox_curso = ttk.Combobox(frame_entrys, textvariable=selected_key,values=list(self.dict_cursos.values()),
+
+        ttk.Label(frame_entrys, text="Curso").grid(row=4, column=0, sticky="W")
+        self.combobox_curso_disciplina = ttk.Combobox(frame_entrys, textvariable=selected_key,values=list(self.dict_cursos.values()),
                                      state="readonly")
-        self.combobox_curso.grid(row=3, column=0, pady=5, padx=10, sticky="W")
+        self.combobox_curso_disciplina.grid(row=5, column=0, pady=5, padx=10, sticky="nsew")
 
 
-        ttk.Button(frame_entrys, text="Adicionar", command=lambda: self.adicionar_disciplina()).grid(row=3, column=1, pady=10)
-        ttk.Button(frame_entrys, text="Filtrar", command=lambda: self.filtrar_disciplina_curso()).grid(row=4, column=1,pady=10)
-        ttk.Button(frame_entrys, text="Editar", command=lambda: self.editar_disciplina()).grid(row=5, column=1, pady=10)
-        ttk.Button(frame_entrys, text="Excluir", command=lambda: self.exluir_disciplina()).grid(row=6, column=1, pady=10)
+
+        ttk.Button(frame_entrys, text="Adicionar", command=self.adicionar_disciplina).grid(row=6, column=0, pady=10)
+        ttk.Button(frame_entrys, text="Filtrar", command=self.filtrar_disciplina_curso).grid(row=6, column=1,pady=10)
+        ttk.Button(frame_entrys, text="Editar", command=self.editar_disciplina).grid(row=7, column=0, pady=10)
+        ttk.Button(frame_entrys, text="Excluir", command=self.exluir_disciplina).grid(row=7, column=1, pady=10)
 
         self.tree_disciplina.bind("<<TreeviewSelect>>", self.on_disciplina_select)
 
@@ -127,10 +137,10 @@ class TelaAdmin(tk.Frame):
         frame_tree = ttk.Frame(self.inscricoes_frame, padding=10)
         frame_tree.grid(row=0, column=1, sticky="NSE")
 
-        colunas = ['ano', 'semestre', 'sim1', 'sim2', 'av', 'avs', 'nf', 'situacao', 'matricula_aluno', 'codigo']
+        colunas = ['id', 'sim1', 'sim2', 'av', 'avs', 'nf', 'situacao', 'Nome Aluno', 'Disciplina']
 
         data = self.inscricao_repository.get_inscricoes()
-        width = int(920 / len(colunas))
+        width = [20, 80, 80, 80, 80, 80, 130, 170, 200]
 
         self.tree_inscricao = Treeview(frame_tree, colunas, data, width)
 
@@ -138,23 +148,44 @@ class TelaAdmin(tk.Frame):
         self.inscricoes_frame.grid_columnconfigure(1, weight=3)
         self.inscricoes_frame.grid_rowconfigure(0, weight=1)
 
-        ttk.Label(frame_entrys, text="SIM1: ").grid(row=0, column=0, sticky="W")
+        selected_key = tk.StringVar()
+        selected_key2 = tk.StringVar()
+
+        ttk.Label(frame_entrys, text="Curso").grid(row=0, column=0, sticky="W")
+        self.combobox_curso_inscricao = ttk.Combobox(frame_entrys, textvariable=selected_key,
+                                           values=list(self.dict_cursos.values()),
+                                           state="readonly")
+        self.combobox_curso_inscricao.grid(row=1, column=0, pady=5, padx=10)
+
+        ttk.Label(frame_entrys, text="SIM1: ").grid(row=4, column=0, sticky="W")
         self.sim1_entry = ttk.Entry(frame_entrys, width=12)
-        self.sim1_entry.grid(row=1, column=0, pady=5, padx=10)
+        self.sim1_entry.grid(row=5, column=0, pady=5, padx=10, sticky="W")
 
-        ttk.Label(frame_entrys, text="SIM2: ").grid(row=0, column=1, sticky="W")
+        ttk.Label(frame_entrys, text="SIM2: ").grid(row=4, column=1, sticky="W")
         self.sim2_entry = ttk.Entry(frame_entrys, width=12)
-        self.sim2_entry.grid(row=1, column=1, pady=5, padx=10)
+        self.sim2_entry.grid(row=5, column=1, pady=5, sticky="W")
 
-        ttk.Label(frame_entrys, text="AV: ").grid(row=2, column=0, sticky="W")
+        ttk.Label(frame_entrys, text="AV: ").grid(row=6, column=0, sticky="W")
         self.av_entry = ttk.Entry(frame_entrys, width=12)
-        self.av_entry.grid(row=3, column=0, pady=5, padx=10)
+        self.av_entry.grid(row=7, column=0, pady=5, padx=10, sticky="W")
 
-        ttk.Label(frame_entrys, text="AVS: ").grid(row=2, column=1, sticky="W")
+        ttk.Label(frame_entrys, text="AVS: ").grid(row=6, column=1, sticky="W")
         self.avs_entry = ttk.Entry(frame_entrys, width=12)
-        self.avs_entry.grid(row=3, column=1, pady=5, padx=10)
+        self.avs_entry.grid(row=7, column=1, pady=5, sticky="W")
 
-        ttk.Button(frame_entrys, text="Adicionar", command=self.adicionar_nota).grid(row=4, column=0, pady=10)
+        ttk.Button(frame_entrys, text="Adicionar", command=self.adicionar_nota).grid(row=8, column=0, pady=10, sticky="W")
+        ttk.Button(frame_entrys, text="Filtrar por curso", command=self.filtrar_inscricao_curso).grid(row=8, column=1, pady=10, sticky="W")
+
+
+        ttk.Label(frame_entrys, text="Disciplina").grid(row=2, column=0, sticky="W")
+        self.combobox_disciplina_inscricao = ttk.Combobox(frame_entrys, textvariable=selected_key2,
+                                                          values=list(self.dict_disciplinas.values()),
+                                                          state="readonly")
+        self.combobox_disciplina_inscricao.grid(row=3, column=0, pady=5, padx=10)
+        ttk.Button(frame_entrys, text="Filtrar por disciplina", command=self.filtrar_inscricao_disciplina).grid(row=9,
+                                                                                                           column=0,
+                                                                                                           pady=10,
+                                                                                                           sticky="W")
 
         self.tree_inscricao.bind('<<TreeviewSelect>>', self.on_inscricao_select)
 
@@ -174,10 +205,14 @@ class TelaAdmin(tk.Frame):
             if values:
                 self.codigo_entry.delete(0, tk.END)
                 self.nome_disciplina_entry.delete(0, tk.END)
-                self.combobox_curso.delete(0, tk.END)
+                self.combobox_curso_disciplina.delete(0, tk.END)
+                self.ano_entry.delete(0, tk.END)
+                self.semestre_entry.delete(0, tk.END)
                 self.codigo_entry.insert(0, values[0])
                 self.nome_disciplina_entry.insert(0, values[1])
-                self.combobox_curso.insert(0, values[2])
+                self.combobox_curso_disciplina.insert(0, values[2])
+                self.ano_entry.insert(0, values[3])
+                self.semestre_entry.insert(0, values[4])
 
     def on_inscricao_select(self, event):
         selected_item = self.tree_inscricao.focus()
@@ -190,10 +225,10 @@ class TelaAdmin(tk.Frame):
                 self.av_entry.delete(0, tk.END)
                 self.avs_entry.delete(0, tk.END)
 
-                self.sim1_entry.insert(0, value[2])
-                self.sim2_entry.insert(0, value[3])
-                self.av_entry.insert(0, value[4])
-                self.avs_entry.insert(0, value[5])
+                self.sim1_entry.insert(0, value[1])
+                self.sim2_entry.insert(0, value[2])
+                self.av_entry.insert(0, value[3])
+                self.avs_entry.insert(0, value[4])
 
     def adicionar_curso(self):
         if len(self.name_curso_entry.get()) > 3:
@@ -209,16 +244,18 @@ class TelaAdmin(tk.Frame):
 
     def adicionar_disciplina(self):
         nome = self.nome_disciplina_entry.get()
-        id_curso = int({v: k for k, v in self.dict_cursos.items()}.get(self.combobox_curso.get()))
+        id_curso = int({v: k for k, v in self.dict_cursos.items()}.get(self.combobox_curso_disciplina.get()))
         codigo = self.codigo_entry.get()
+        ano = self.ano_entry.get() or int(datetime.now().strftime("%Y"))
+        semestre = self.semestre_entry.get() or 1 if int(datetime.now().strftime("%m")) <= 6 else 2
 
         if len(nome) > 7 and id_curso and len(codigo) > 3:
-            self.disciplina_repository.adicionar_disciplina(Disciplina(codigo, nome, id_curso))
+            self.disciplina_repository.adicionar_disciplina(Disciplina(codigo, nome, id_curso, ano, semestre))
             data = self.disciplina_repository.get_disciplinas()
             self.tree_disciplina.atualizar(data)
             self.atualizar()
             self.nome_disciplina_entry.delete(0, tk.END)
-            self.combobox_curso.delete(0, tk.END)
+            self.combobox_curso_disciplina.delete(0, tk.END)
             self.codigo_entry.delete(0, tk.END)
             messagebox.showinfo("SUCESSO!", "Disciplina adicionada!")
             return
@@ -230,14 +267,14 @@ class TelaAdmin(tk.Frame):
         if selected_item:
             values = self.tree_inscricao.item(selected_item, 'values')
 
-            sim1 = float(self.tratar_valor(self.sim1_entry.get()))
-            sim2 = float(self.tratar_valor(self.sim2_entry.get()))
-            av = float(self.tratar_valor(self.av_entry.get()))
-            avs = float(self.tratar_valor(self.avs_entry.get()))
+            sim1 = float(self.sim1_entry.get()) if self.sim1_entry.get() != "" else 0
+            sim2 = float(self.sim2_entry.get()) if self.sim2_entry.get() != "" else 0
+            av = float(self.av_entry.get()) if self.av_entry.get() != "" else 0
+            avs = float(self.avs_entry.get()) if self.avs_entry.get() != "" else 0
 
             nota = Nota(sim1, sim2, av, avs)
 
-            self.inscricao_repository.insert_nota(nota, values[8], values[9])
+            self.inscricao_repository.insert_nota(nota, values[0])
             data = self.inscricao_repository.get_inscricoes()
             self.tree_inscricao.atualizar(data)
             self.atualizar()
@@ -246,10 +283,6 @@ class TelaAdmin(tk.Frame):
             self.sim2_entry.delete(0, tk.END)
             self.av_entry.delete(0, tk.END)
             self.avs_entry.delete(0, tk.END)
-
-
-
-
 
     def editar_curso(self):
         selected_item = self.tree_curso.focus()
@@ -271,7 +304,9 @@ class TelaAdmin(tk.Frame):
         if selected_item:
             values = self.tree_disciplina.item(selected_item, 'values')
             nome = self.nome_disciplina_entry.get()
-            self.disciplina_repository.editar(values[0], nome)
+            ano = self.ano_entry.get() or int(datetime.now().strftime("%Y"))
+            semestre = self.semestre_entry.get() or 1 if int(datetime.now().strftime("%m")) <= 6 else 2
+            self.disciplina_repository.editar(values[0], nome, ano, semestre)
             self.nome_disciplina_entry.delete(0, tk.END)
             data = self.disciplina_repository.get_disciplinas()
             self.tree_disciplina.atualizar(data)
@@ -297,12 +332,15 @@ class TelaAdmin(tk.Frame):
         selected_item = self.tree_curso.focus()
         if selected_item:
             values = self.tree_curso.item(selected_item, 'values')
-            self.curso_repository.remove(values[0])
-            data = self.curso_repository.get_cursos()
-            self.tree_curso.atualizar(data)
-            self.atualizar()
-            messagebox.showinfo("SUCESSO!", "Curso removido!")
-            return
+            resposta = messagebox.askyesno("AVISO!", f"Curso você deseja realmente remover {values[1]}?")
+            if resposta:
+                self.curso_repository.remove(values[0])
+                data = self.curso_repository.get_cursos()
+                self.tree_curso.atualizar(data)
+                self.atualizar()
+                return
+            else:
+                return
         messagebox.showerror("ERRO!", "Não foi possível remover o curso")
 
     def atualizar(self):
@@ -312,7 +350,7 @@ class TelaAdmin(tk.Frame):
 
     def filtrar_disciplina_curso(self):
         try:
-            id_curso = int({v: k for k, v in self.dict_cursos.items()}.get(self.combobox_curso.get()))
+            id_curso = int({v: k for k, v in self.dict_cursos.items()}.get(self.combobox_curso_disciplina.get()))
 
             if id_curso:
                 data = self.disciplina_repository.get_disciplinas_by_curso(id_curso)
@@ -324,10 +362,43 @@ class TelaAdmin(tk.Frame):
             self.atualizar()
             self.tree_disciplina.atualizar(data)
 
-    def tratar_valor(self, value):
-        return value if value != "" else None
+    def filtrar_inscricao_curso(self):
+        try:
+            id_curso = int({v: k for k, v in self.dict_cursos.items()}.get(self.combobox_curso_inscricao.get()))
+
+            if id_curso:
+                data = self.inscricao_repository.get_inscricoes_by_curso(id_curso)
+                self.clicked_filter = True
+                self.dict_disciplinas = self.adiciona_dados_combobox_disciplinas(id_curso)
+                self.atualizar()
+                self.tree_inscricao.atualizar(data)
+                return
+        except Exception as e:
+            data = self.inscricao_repository.get_inscricoes()
+            self.dict_disciplinas = dict(self.disciplina_repository.get_all_to_dict())
+            self.atualizar()
+            self.tree_inscricao.atualizar(data)
+
+    def filtrar_inscricao_disciplina(self):
+        try:
+            codigo = {v: k for k, v in self.dict_disciplinas.items()}.get(self.combobox_disciplina_inscricao.get())
+
+            if codigo:
+                data = self.inscricao_repository.get_inscricoes_by_disciplina(codigo)
+                self.atualizar()
+                self.tree_inscricao.atualizar(data)
+                return
+        except Exception as e:
+            data = self.inscricao_repository.get_inscricoes()
+            self.dict_disciplinas = dict(self.disciplina_repository.get_all_to_dict())
+            self.atualizar()
+            self.tree_inscricao.atualizar(data)
+
+    def adiciona_dados_combobox_disciplinas(self, id_curso):
+        return dict(self.disciplina_repository.get_by_curso_id(id_curso))
 
     def deslogar(self):
        from Telas.TelaLogin import TelaLogin
+       self.atualizar()
        self.controller.geometry("300x200")
        self.controller.mostrar_tela(TelaLogin)
